+++ static/app/utils/replays/extractDomNodes.tsx
