+++ static/app/utils/replays/hooks/useExtractDomNodes.tsx
@@ -1,5 +1,5 @@
 import {useQuery, type UseQueryResult} from 'sentry/utils/queryClient';
-import type {Extraction} from 'sentry/utils/replays/extractHtml';
+import type {Extraction} from 'sentry/utils/replays/extractDomNodes';
 import type ReplayReader from 'sentry/utils/replays/replayReader';
 import type {ReplayFrame} from 'sentry/utils/replays/types';
